♻️ Move other endpoints to own blueprints
